feat: :sparkles: user managing

+++ app/handlers/admin.py
@@ -21,30 +21,53 @@ async def get_admin_dashboard(message: types.Message):
 async def make_role(message: types.Message, command: CommandObject): # /make userid.role
     if command.args:
         userid, role = command.args.split('.')
-        try:
-            result = await add_user(int(userid), role)
-            print(result, file=sys.stderr)
-            users = await get_users_by_role(role)
-            print(f"Users by role {role}: {users}", file=sys.stderr)
-            await message.answer(f"Пользователь {await get_username(userid)} добавлен в группу {role}.") #защита на введение admin, ролей которых не существует
-        except Exception as e:
-            await message.answer(f"Exception found: {e}")
-    else:
-        pass
+        if role in ["worker", "spectator"]:
+            try:
+                result = await add_user(int(userid), role)
+                print(result, file=sys.stderr)
+                await message.answer(f"Пользователь {await get_username(userid)} добавлен в группу {role}.") #защита на введение admin, ролей которых не существует
+            except Exception as e:
+                await message.answer(f"Exception found: {e}")
+        else:
+            await message.answer(f"Вы не можете добавить пользователя с ролью {role}. Доступные роли: worker, spectator.")
 
 @router.message(Command("rm"), F.from_user.id.in_(admin))
 async def remove_role(message: types.Message, command: CommandObject): # /rm userid
     if command.args:
         userid = int(command.args)
+        role = await custom_sql(f"SELECT role FROM users WHERE userid = {userid}", fetchval=True)
+        if role!="admin":
+            try:
+                result = await delete_user(userid)
+                print(result, file=sys.stderr)
+                await message.answer(f"Пользователь {await get_username(userid)} успешно удалён.")
+            except Exception as e:
+                await message.answer(f"Exception found: {e}")
+        else:
+            await message.answer(f"Нет прав для удаления пользователя {get_username(userid)}, поскольку этот пользователь {role}.")
+
+@router.message(Command("makeadmin"), F.from_user.id==ROOT)
+async def make_admin(message: types.Message, command: CommandObject): # /make userid.role
+    if command.args:
+        userid = int(command.args)
         try:
-            result = await delete_user(userid)
+            result = await add_user(userid, "admin")
             print(result, file=sys.stderr)
-            await message.answer(f"Пользователь {await get_username(userid)} успешно удалён.")
+            await message.answer(f"Пользователь {await get_username(userid)} добавлен в список администраторов.")
         except Exception as e:
             await message.answer(f"Exception found: {e}")
-    else:
-        pass
 
-#@router.message(Command("makeadmin"), F.from_user.id==ROOT)
-
-#@router.message(Command("rmadmin"), F.from_user.id==ROOT)
+@router.message(Command("rmadmin"), F.from_user.id==ROOT)
+async def remove_admin(message: types.Message, command: CommandObject):
+    if command.args:
+        userid = int(command.args)
+        role = await custom_sql(f"SELECT role FROM users WHERE userid = {userid}", fetchval=True)
+        if role=="admin":
+            try:
+                result = await delete_user(userid)
+                print(result, file=sys.stderr)
+                await message.answer(f"Пользователь {await get_username(userid)} удалён из списка администраторов.")
+            except Exception as e:
+                await message.answer(f"Exception found: {e}")
+        else:
+            await message.answer(f"Пользователь {await get_username(userid)} не администратор, а {role}.")
